refactor: AdminControllerTest에 있던 예약 테스트 ReservationControllerTest로 이동
